v3/requirements.txt
# VideoMAE Baseline Requirements
# Python 3.8+ required

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Video processing
pytorchvideo>=0.1.5
av>=10.0.0
opencv-python>=4.8.0

# Image processing
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# Evaluation metrics
scikit-image>=0.21.0
torchmetrics>=1.0.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0

# Optional but recommended
accelerate>=0.20.0  # For faster training
einops>=0.7.0       # For tensor operations